Backend/HeartOFProject/resume-skill-extractor/requirements.txt
pdfplumber
python-docx
spacy
fastapi
uvicorn
python-multipart
python-dotenv
requests
groq
